server/requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.1.1
django-filter==24.3
djangorestframework==3.15.2
idna==3.10
kombu==5.4.2
Markdown==3.7
numpy==2.1.1
pandas==2.2.3
prompt_toolkit==3.0.48
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
pytz==2024.2
redis==5.1.0
requests==2.32.3
six==1.16.0
SQLAlchemy==2.0.35
sqlparse==0.5.1
tomli==2.0.1
types-PyYAML==6.0.12.20240917
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
